requirements.txt
torch
torchaudio
transformers
dataclasses_json
tqdm
evaluate
scikit-learn
ffmpeg-python
wandb
google-genai